update bar text
